+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/PauseMenu.java
@@ -57,6 +57,7 @@ public class PauseMenu extends Application {
   private boolean alreadyMuted = false;
 
   private final List<FrontendScreenAsset> screenAssets = new ArrayList<>();
+  private final List<PauseMenuStatusChangeListener> listeners = new ArrayList<>();
 
   private static PauseMenu INSTANCE = null;
   private MenuController menuController;
@@ -103,6 +104,10 @@ public class PauseMenu extends Application {
 
   public boolean isVisible() {
     return visible;
+  }
+
+  public void addListener(PauseMenuStatusChangeListener listener) {
+    this.listeners.add(listener);
   }
 
   public void loadPauseMenu() {
@@ -295,6 +300,8 @@ public class PauseMenu extends Application {
           if (pauseMenuSettings.isMuteOnPause() && !alreadyMuted) {
             NirCmd.muteSystem(true);
           }
+
+          listeners.forEach(PauseMenuStatusChangeListener::pauseMenuShow);
           return true;
         }).thenAcceptLater((result) -> {
           long start = System.currentTimeMillis();
@@ -344,6 +351,7 @@ public class PauseMenu extends Application {
         asset.getFrontendScreenController().dispose();
       });
       screenAssets.clear();
+      listeners.forEach(PauseMenuStatusChangeListener::pauseMenuHide);
     });